pybind/cephfs: don't attempt to unlink root in rmtree
Don't attempt to unlink '/' as part of recursive delete when rmtree('/')
is called as it is an invalid op.

Fixes: https://tracker.ceph.com/issues/75924
(cherry picked from commit defbc0651e44b00060cdcf11b28c714f189b1c72)

+++ src/pybind/cephfs/cephfs.pyx
@@ -3419,6 +3419,9 @@ class RmtreeDir:
         and tell caller whether to continue or break loop based through the
         return value.
         '''
+        if self.name == b'/':
+            return
+
         try:
             self.fs.unlinkat(self.parent_dir_fd, self.name, AT_REMOVEDIR)
